Use Guid on delete and set active state

+++ src/MFO.CatalogService.Infrastructure/Repositories/BrandRepository.cs
@@ -34,8 +34,14 @@ public class BrandRepository : IBrandRepository
         return brand;
     }
 
-    public async Task<bool> DeleteBrandAsync(Brand brand, CancellationToken cancellationToken)
+    public async Task<bool> DeleteBrandAsync(Guid brandId, CancellationToken cancellationToken)
     {
+        var brand = await _db.Brands.FindAsync([brandId], cancellationToken);
+        if (brand is null)
+        {
+            return false;
+        }
+
         _db.Brands.Remove(brand);
         await _db.SaveChangesAsync(cancellationToken);
         return true;
